requirements.txt
# Core dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-experimental>=0.0.47
langchain-openai>=0.0.5

# Database
pymysql>=1.1.0
sqlalchemy>=2.0.0

# Vector store and embeddings
chromadb>=0.4.0

# Environment and utilities
python-dotenv>=1.0.0

# OpenAI
openai>=1.0.0
